wip: split up timer store and simplify
